# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Scorekeeper API
  description: |
    API for the Scorekeeper word puzzle game.

    See also: [Planned Features](https://hannahscovill.github.io/scorekeeper/planned)
  version: 0.1.0
servers:
  - url: https://api.wordles.dev
    description: Remote
  - url: http://localhost:8080
    description: Local
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /issues:
    post:
      summary: Submit an issue report
      description: |
        Create a GitHub issue on behalf of an authenticated user.

        The reporter's user ID is extracted from the JWT and included in the
        issue body. Display name is resolved from the Auth0 user profile when
        available. PostHog session ID is included for session replay correlation.

        The issue body includes separate Client and Server environment sections.
        Client metadata (browser, URL, environment name, commit hash) is sent
        by the frontend. Server metadata (environment name, commit hash) is
        appended by the backend from its own deployment context.

        Includes Turnstile CAPTCHA verification, honeypot spam detection,
        and per-IP rate limiting (5 issues per hour).
      operationId: createIssue
      tags:
        - Issues
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "schemas.yaml#/IssueRequest"
            examples:
              bug:
                $ref: "data-examples/examples.yaml#/IssueRequestBug"
              feature:
                $ref: "data-examples/examples.yaml#/IssueRequestFeature"
      responses:
        "201":
          description: Issue created successfully
          content:
            application/json:
              schema:
                $ref: "schemas.yaml#/IssueReportResponse"
              examples:
                default:
                  $ref: "data-examples/examples.yaml#/IssueReportResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "schemas.yaml#/ErrorResponse"
              example:
                message: "Server configuration error"
...
  /profile/mobile-test-track-signup:
    post:
      summary: Opt into the internal mobile test track (authenticated)
      description: |
        Opts the authenticated user into the iOS and/or Android internal test
        track, writing directly to their Auth0 user_metadata.

        Idempotent per platform — resubmitting a platform that's already
        opted in is a no-op (no Auth0 write, no GitHub notification).

        Notifies via a GitHub issue (label `mobile-test-user-opt-in`)
        containing the user's display name and Auth0 user ID only — the
        email is never included, even though it's a field on the account.
        GitHub notification is best-effort: if GitHub credentials aren't
        configured, the opt-in still succeeds and the notification is
        silently skipped.

        **No Turnstile token required** — a valid JWT already proves the
        request is from a real logged-in user, so CAPTCHA verification
        (required on the anonymous `POST /mobile-test-track-signup` below) would be
        redundant here.
      operationId: mobileTestTrackSignup
      tags:
        - Mobile Test Track Signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "schemas.yaml#/MobileTestTrackSignupRequest"
            examples:
              default:
                $ref: "data-examples/examples.yaml#/MobileTestTrackSignupRequest"
      responses:
        "200":
          description: Current opt-in state after the call
          content:
            application/json:
              schema:
                $ref: "schemas.yaml#/MobileTestTrackSignupResponse"
              examples:
                default:
                  $ref: "data-examples/examples.yaml#/MobileTestTrackSignupResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /mobile-test-track-signup:
    post:
      summary: Opt into the internal mobile test track (anonymous)
      description: |
        Opts an anonymous visitor into the iOS and/or Android internal test
        track by email, without requiring an account. Creates (or finds) a
        passwordless Auth0 "email" connection lead and writes the same
        opt-in fields used by `POST /profile/mobile-test-track-signup`.

        Idempotent per email/platform — resubmitting a platform already
        opted into is a no-op. Notifies via the same GitHub issue mechanism
        as the authenticated endpoint — the issue includes only the Auth0
        user ID, **never the email or any hint of it**.

        Public and unauthenticated, so it's gated the same way `/issues`
        is: Turnstile CAPTCHA verification (required, not best-effort),
        per-IP rate limiting (5 requests per hour, separate quota from
        `/issues`), and a honeypot field.
      operationId: anonymousMobileTestTrackSignup
      tags:
        - Mobile Test Track Signup
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "schemas.yaml#/AnonymousMobileTestTrackSignupRequest"
            examples:
              default:
                $ref: "data-examples/examples.yaml#/AnonymousMobileTestTrackSignupRequest"
      responses:
        "200":
          description: Current opt-in state after the call
          content:
            application/json:
              schema:
                $ref: "schemas.yaml#/MobileTestTrackSignupResponse"
              examples:
                default:
                  $ref: "data-examples/examples.yaml#/MobileTestTrackSignupResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token in Authorization header
    cookieAuth:
      type: apiKey
      in: cookie
      name: wordle_session
      description: JWT token in wordle_session cookie
  responses:
...
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "schemas.yaml#/ErrorResponse"
          example:
            message: "bad request"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "schemas.yaml#/ErrorResponse"
          example:
            message: "unauthorized"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "schemas.yaml#/ErrorResponse"
          example:
            message: "forbidden"
...
    TooManyRequests:
      description: Too Many Requests
      content:
        application/json:
          schema:
            $ref: "schemas.yaml#/ErrorResponse"
          example:
            message: "Rate limit exceeded. Please try again later."
...